requirements.txt
pandas
joblib
spotipy
selenium
python-dotenv
argparse
typer